inference.py
import os
import json
from openai import OpenAI

# Required variables as per Hackathon specs
API_BASE_URL = os.getenv("API_BASE_URL", "https://router.huggingface.co/v1")
MODEL_NAME = os.getenv("MODEL_NAME", "Qwen/Qwen3-1.7B-Instruct")
HF_TOKEN = os.getenv("HF_TOKEN")

# Initialize OpenAI Client using HF Router
client = OpenAI(base_url=API_BASE_URL, api_key=HF_TOKEN)

def run_evaluation():
    print("[START] - Starting Healthcare Triage Evaluation")
    
    # These prompts test your environment logic for 'prioritize' and 'extract_vitals'
    prompts = [
        "The patient has chest pain. Return only the word 'urgent'.",
        "Extract only the number from: 'Vitals: Temp 103F'."
    ]
    
    total_reward = 0.0
    
    for i, prompt in enumerate(prompts):
        try:
            response = client.chat.completions.create(
                model=MODEL_NAME,
                messages=[{"role": "user", "content": prompt}],
                max_tokens=50
            )
            
            answer = response.choices[0].message.content.strip()
            
            # Logic to simulate rewards based on your HealthTriageEnvironment logic
            # Reward 0.5 for 'urgent' (Task 1) and 0.5 for '103' (Task 2)
            step_reward = 0.0
            if "urgent" in answer.lower() or "103" in answer:
                step_reward = 0.5
            
            total_reward += step_reward
            
            # MANDATORY LOGGING FORMAT for Scaler Grader
            print(f"[STEP] - Step: {i+1} | Action: {answer} | Reward: {step_reward}")
        except Exception as e:
            print(f"[STEP] - Step: {i+1} | Error: {str(e)} | Reward: 0.0")

    final_score = total_reward / len(prompts)
    print(f"[END] - Final Score: {final_score}")

if __name__ == "__main__":
    if not HF_TOKEN:
        print("ERROR: HF_TOKEN environment variable is not set in Secrets.")
    else:
        run_evaluation()